+++ packages/curvenote/src/submissions/submit.ts
@@ -21,85 +21,19 @@ import {
 } from './utils.js';
 import inquirer from 'inquirer';
 import { uploadContentAndDeployToPrivateCdn } from '../utils/web.js';
-
-type SubmitOpts = { kind?: string; yes?: boolean; info: boolean; draft?: boolean };
-
-function kindQuestions(kinds: { name: string }[]) {
-  return {
-    name: 'kinds',
-    type: 'list',
-    message: 'What kind of submission are you making?',
-    choices: kinds.map(({ name }) => ({ name, value: name })),
-  };
-}
-
-function venueQuestion(session: ISession) {
-  return {
-    name: 'venue',
-    type: 'input',
-    message: 'Enter the venue name you want to submit to?',
-    filter: (venue: string) => venue.toLowerCase(),
-    validate: async (venue: string) => {
-      if (venue.length < 3) {
-        return 'Venue name must be at least 3 characters';
-      }
-      try {
-        await getFromJournals(session, `sites/${venue}`);
-      } catch (err) {
-        return `Venue "${venue}" not found.`;
-      }
-      return true;
-    },
-  };
-}
-
-async function determineSubmissionKind(session: ISession, venue: string, opts?: { kind?: string }) {
-  let kinds;
-  try {
-    kinds = await getFromJournals(session, `sites/${venue}/kinds`);
-  } catch (err: any) {
-    session.log.info(
-      `${chalk.red(`🚨 could not get submission kinds listing from venue ${venue}`)}`,
-    );
-    process.exit(1);
-  }
-
-  let kind;
-  if (opts?.kind) {
-    if (
-      !kinds.items
-        .map(({ name }: { name: string }) => name.toLowerCase())
-        .includes(opts.kind.toLowerCase())
-    ) {
-      session.log.info(
-        `${chalk.red(`🚨 submission kind "${opts.kind}" is not accepted at venue ${venue}`)}`,
-      );
-      process.exit(1);
-    }
-    kind = opts?.kind;
-  } else if (kinds.items.length === 1) {
-    kind = kinds.items[0].name;
-  } else {
-    const response = await inquirer.prompt([kindQuestions(kinds.items)]);
-    kind = response.kinds;
-  }
-  return kind;
-}
-
-async function performCleanRebuild(session: ISession, opts?: SubmitOpts) {
-  session.log.info('\n\n\t✨✨✨  performing a clean re-build of your work  ✨✨✨\n\n');
-  await clean(session, [], { site: true, html: true, temp: true, exports: true, yes: true });
-  const exportOptionsList = await collectAllBuildExportOptions(session, [], { all: true });
-  const exportLogList = exportOptionsList.map((exportOptions) => {
-    return `${path.relative('.', exportOptions.$file)} -> ${exportOptions.output}`;
-  });
-  session.log.info(`📬 Performing exports:\n   ${exportLogList.join('\n   ')}`);
-  await localArticleExport(session, exportOptionsList, {});
-  session.log.info(`⛴ Exports complete`);
-  // Build the files in the content folder and process them
-  await buildSite(session, addOxaTransformersToOpts(session, opts ?? {}));
-  session.log.info(`✅ Work rebuild complete`);
-}
+import {
+  ensureVenue,
+  getSiteConfig,
+  checkVenueExists,
+  checkVenueAccess,
+  determineSubmissionKind,
+  performCleanRebuild,
+  celebrate,
+  SubmitOpts,
+  confirmUpdateToExistingSubmission,
+  updateExistingSubmission,
+  getTransferData,
+} from './submit.utils.js';
 
 export async function submit(session: ISession, venue: string, opts?: SubmitOpts) {
   if (session.isAnon) {
@@ -112,118 +46,36 @@ export async function submit(session: ISession, venue: string, opts?: SubmitOpts
   // TODO check the venue allows for submissions & updates to the submission
   // TODO check user has permission to submit /  update a submission
 
-  const siteConfig = selectors.selectCurrentSiteConfig(session.store.getState());
-  if (!siteConfig) {
-    throw new Error('🧐 No site config found.');
-  }
-
-  session.log.debug('Checking for a "transfer.yml" file...');
-  const transferData = await loadTransferFile(session);
-  if (transferData) {
-    session.log.info(`${chalk.bold(`🧐 Found a "transfer.yml" in this folder.`)}`);
-  }
-
-  if (!venue) {
-    session.log.debug('No venue provided, prompting user...');
-    const answer = await inquirer.prompt([venueQuestion(session)]);
-    venue = answer.venue;
-    session.log.debug(`user answered: "${venue}"`);
-  }
-
-  // check site exists and can make a submission
-  try {
-    session.log.debug(`GET from journals API sites/${venue}`);
-    await getFromJournals(session, `sites/${venue}`);
-    session.log.debug(`found venue "${venue}"`);
-  } catch (err) {
-    session.log.debug(err);
-    session.log.error(`${chalk.red(`👩🏻‍🔬 venue "${venue}" not found.`)}`);
-    process.exit(1);
-  }
-
-  // PREFLIGHT CHECK
-  try {
-    await getFromJournals(session, `sites/${venue}/access`);
-    session.log.info(`${chalk.green(`👩🏻‍🔬 venue "${venue}" is accepting submissions.`)}`);
-  } catch (err) {
-    session.log.info(`${chalk.red(`🚦 venue "${venue}" is not accepting submissions.`)}`);
-    process.exit(1);
-  }
+  const siteConfig = getSiteConfig(session);
+  let transferData = await getTransferData(session, opts);
+  venue = await ensureVenue(session, venue);
+  await checkVenueExists(session, venue);
+  await checkVenueAccess(session, venue);
 
   //
   // Options, checks and prompts
   //
   let kind: string | undefined;
-  if (transferData?.[venue]) {
-    session.log.debug('found venue in transfer.yml, existing submission');
-    const lastSubDate = transferData[venue].submissionVersion?.date_created;
-    session.log.info(
-      chalk.bold(
-        `🗓 you last submitted this work to "${venue}" on ${
-          lastSubDate ? format(new Date(lastSubDate), 'dd MMM, yyyy HH:mm:ss') : 'unknown'
-        }.`,
-      ),
-    );
-
-    session.log.info(`📡 Checking submission status...`);
-
-    let existingSubmission;
-    try {
-      session.log.debug(
-        `GET from journals API sites/${venue}/submissions/${transferData[venue].submission?.id}`,
-      );
-      existingSubmission = await getFromJournals(
-        session,
-        `sites/${venue}/submissions/${transferData[venue].submission?.id}`,
-      );
-    } catch (err: any) {
-      session.log.debug(err);
-      session.log.info(
-        `${chalk.red(`🚨 submission not found, or you do not have permission to update it`)}`,
-      );
-      process.exit(1);
-    }
-    session.log.info(
-      `✅ Submission found, ${existingSubmission?.versions.length} version${
-        existingSubmission?.versions.length > 1 ? 's' : ''
-      } present, latest status: ${existingSubmission?.versions[0].status}.`,
-    );
-
-    try {
-      session.log.debug(`GET from journals API my/works/${transferData[venue].work?.id}`);
-      await getFromJournals(session, `my/works/${transferData[venue].work?.id}`);
-    } catch (err) {
-      session.log.debug(err);
-      session.log.info(
-        `${chalk.red(
-          `🚨 the work related to your submission was not found, or you do not have permission to update it`,
-        )}`,
-      );
-      process.exit(1);
-    }
-
-    if (opts?.kind) {
-      session.log.info(
-        `🪧  NOTE: the --kind option was provided, but will be ignored as you are updating an existing submission`,
-      );
-    }
-    kind = existingSubmission?.kind;
-    session.log.debug(`resolved kind to ${kind}`);
-
-    await confirmOrExit(
-      `Update your submission to "${venue}" based on the contents of your local folder?`,
-      opts,
-    );
+  if (transferData?.[venue] && !opts?.draft) {
+    kind = await confirmUpdateToExistingSubmission(session, venue, transferData[venue], opts);
   } else {
     //
     // NEW SUBMISSIONS
     //
     session.log.debug('new submission');
     kind = await determineSubmissionKind(session, venue, opts);
-    session.log.debug(`resolved kind to ${kind}`);
+
+    if (opts?.draft)
+      session.log.info(
+        `📝 ${chalk.bold.yellow(
+          `Making a draft submission, this is just for you and will not be processed as a submission by "${venue}".`,
+        )}`,
+      );
 
     await confirmOrExit(
-      `Start a new submission to "${venue}" based on the contents of your local folder?`,
+      opts?.draft
+        ? `Submit your draft to "${venue}" based on the contents of your local folder?`
+        : `Start a new submission to "${venue}" based on the contents of your local folder?`,
       opts,
     );
   }
@@ -232,61 +84,19 @@ export async function submit(session: ISession, venue: string, opts?: SubmitOpts
   // Process local folder and upload stuff
   //
   await performCleanRebuild(session, opts);
-  const cdnKey = await uploadContentAndDeployToPrivateCdn(session, {
-    ...opts,
-    ci: opts?.yes,
-  });
+  celebrate(session, 'Successfully built your work!');
+  const cdnKey = '96b95ed0-d19d-4c54-b5d9-d10fb7b3d9da';
+  // const cdnKey = await uploadContentAndDeployToPrivateCdn(session, {
+  //   ...opts,
+  //   ci: opts?.yes,
+  // });
   session.log.info(`🚀 ${chalk.bold.green(`Content uploaded with key ${cdnKey}`)}.`);
 
   //
   // Create work and submission
   //
   if (transferData?.[venue] && !opts?.draft) {
-    session.log.debug(`existing submission - upload & post`);
-    const workId = transferData[venue].work?.id;
-    const submissionId = transferData[venue].submission?.id;
-    if (!workId) {
-      session.log.error('🚨 No work id found - invalid transfer.yml');
-      process.exit(1);
-    }
-    if (!submissionId) {
-      session.log.error('🚨 No submission id found - invalid transfer.yml');
-      process.exit(1);
-    }
-
-    try {
-      session.log.debug(`posting new work version...`);
-      const { workVersion } = await postNewWorkVersion(
-        session,
-        workId,
-        cdnKey,
-        session.PRIVATE_CDN,
-      );
-      session.log.debug(`work version posted with id ${workVersion.id}`);
-
-      session.log.debug(`posting new version to existing submission...`);
-      const { submissionVersion } = await postUpdateSubmissionWorkVersion(
-        session,
-        venue,
-        submissionId,
-        workVersion.id,
-      );
-      session.log.debug(`submission version posted with id ${submissionVersion.id}`);
-
-      session.log.info(
-        `🚀 ${chalk.bold.green(`Your submission was successfully updated at "${venue}"`)}.`,
-      );
-
-      await upwriteTransferFile(session, venue, { workVersion, submissionVersion });
-
-      session.log.info(
-        `The "./transfer.yml" file has been updated your submission information, please keep this file or commit this change.`,
-      );
-    } catch (err: any) {
-      session.log.info(`\n\n🚨 ${chalk.bold.red('Could not update your submission')}.`);
-      session.log.info(`📣 ${chalk.red(err.message)}.`);
-      process.exit(1);
-    }
+    await updateExistingSubmission(session, venue, cdnKey, transferData[venue]);
   } else {
     session.log.debug(`new submission - upload & post`);
 
@@ -318,15 +128,15 @@ export async function submit(session: ISession, venue: string, opts?: SubmitOpts
       );
       session.log.debug(`new submission posted with id ${submission.id}`);
 
-      session.log.info(
-        `🚀 ${chalk.bold.green(`Your work was successfully submitted to "${venue}"`)}.`,
-      );
-      if (opts?.draft)
+      if (opts?.draft) {
         session.log.info(
-          `✍️ ${chalk.bold.yellow(
-            `As your work is marked as draft, "${venue}" will not process it as a submission.`,
-          )}.`,
+          `✍️ ${chalk.bold.green(`🚀 Your draft was successfully submitted to "${venue}".`)}.`,
         );
+      } else {
+        session.log.info(
+          `🚀 ${chalk.bold.green(`Your work was successfully submitted to "${venue}"`)}.`,
+        );
+      }
 
       if (opts?.draft) {
         session.log.debug(`generating link for draft submision...`);
@@ -348,9 +158,9 @@ export async function submit(session: ISession, venue: string, opts?: SubmitOpts
           },
         );
 
-        const buildUrl = `${session.JOURNALS_URL}build/${job.id}`;
-
-        session.log.info(chalk.bold(`🔗 access your submission here: ${buildUrl}`));
+        const buildUrl = `${session.JOURNALS_URL.replace('v1/', '')}build/${job.id}`;
+        session.log.info(chalk.bold.green(`🔗 access the build report and draft submission here:`));
+        celebrate(session, buildUrl);
       } else {
         session.log.debug(`writing to transfer.yml...`);
         await upwriteTransferFile(session, venue, {
